# Simplify the API to `@precompile_module(M)`