--- a/Project.toml
+++ b/Project.toml
@@ -1,7 +1,7 @@
 name = "PrecompileSignatures"
 uuid = "91cefc8d-f054-46dc-8f8c-26e11d7c5411"
 authors = ["Rik Huijzer <t.h.huijzer@rug.nl>"]
-version = "1.0.1"
+version = "1.1.0"
 
 [deps]
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"