--- a/Project.toml
+++ b/Project.toml
@@ -1,15 +1,13 @@
 name = "PrecompileSignatures"
 uuid = "91cefc8d-f054-46dc-8f8c-26e11d7c5411"
 authors = ["Rik Huijzer <t.h.huijzer@rug.nl>"]
-version = "1.1.0"
+version = "2.0.0"
 
 [deps]
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
-Scratch = "6c6a2e73-6563-6170-7368-637461726353"
 
 [compat]
 Documenter = "0.27"
-Scratch = "1.1"
 julia = "1.6"
 
 [extras]