--- a/README.md
+++ b/README.md
@@ -15,11 +15,15 @@ pkg> activate Foo
 Next, add the following somewhere in your code:
 
 ```julia
-using PrecompileSignatures: precompile_directives
+module Foo
 
-if ccall(:jl_generating_output, Cint, ()) == 1
-    include(precompile_directives(Foo))
-end
+using PrecompileSignatures: precompile_signatures
+
+[...]
+
+precompile_signatures(Foo)
+
+end # module
 ```
 
 This will generate extra `precompile` directives during the precompilation phase and `include` the generated file.