This PR:

- simplifies the API since package authors now only have to write `@precompile_module(Foo)`.
- avoids writing to a file (not that this saves time tho xD see below)
- fixes the following risks for failures:

1. An include from `include(precompile_directives(M))` turns out to be wrong. Developers wouldn't push this, so the risk doesn't lie there. However, this package could make a mistake in generating precompile directives causing breaking downstream packages? Would be better to show warnings in that case and still allow the package to be used.
2. Other weird things on very weird setups? Probably best is to wrap every `precompile` call in a function with a try-catch.

## Benchmark

On Julia 1.7.3 with Pluto

### Before this PR (via `include(precompile_directives(Pluto))`)

Precompiled in 17 seconds
```
 Section                        ncalls     time    %tot     alloc    %tot
 ────────────────────────────────────────────────────────────────────────
 import Pluto                        1    1.27s    3.9%    160MiB    5.9%
 compiletimes.jl                     1    31.6s   96.1%   2.49GiB   94.1%
   PlutoRunner.run_expression        1    1.75s    5.3%    137MiB    5.1%
   SessionActions.open               1    23.0s   69.9%   1.80GiB   67.9%
   Pluto.run                         1    1.43s    4.3%    110MiB    4.1%
 ────────────────────────────────────────────────────────────────────────
```

### After this PR (via `@precompile_module(Pluto)`)

Precompiled in 17 seconds
```
 Section                        ncalls     time    %tot     alloc    %tot
 ────────────────────────────────────────────────────────────────────────
 import Pluto                        1    1.28s    3.9%    157MiB    5.8%
 compiletimes.jl                     1    31.1s   96.1%   2.49GiB   94.2%
   PlutoRunner.run_expression        1    1.74s    5.4%    137MiB    5.0%
   SessionActions.open               1    22.6s   69.8%   1.80GiB   67.9%
   Pluto.run                         1    1.43s    4.4%    110MiB    4.1%
 ────────────────────────────────────────────────────────────────────────
```